requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0

# OAuth2 & Authentication
django-oauth-toolkit==2.3.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# CORS Support
django-cors-headers==4.3.1

# Database
# psycopg2-binary==2.9.9  # Optional: Nur für PostgreSQL

# Environment Variables
python-decouple==3.8

# Security
cryptography==41.0.7

# Validation
django-phonenumber-field==7.3.0
phonenumbers==8.13.27

# API Documentation
drf-spectacular==0.27.0

# Social Authentication
django-allauth==0.63.3

# Redis for caching/sessions (Optional)
# redis==5.0.1
# django-redis==5.4.0

# Development
python-dotenv==1.0.0
